+++ packages/jest-runtime/src/index.ts
@@ -133,7 +133,9 @@ export default class Runtime {
     this._config = config;
     this._coverageOptions = coverageOptions;
     this._environment = environment;
-    this.registries = new ModuleRegistries();
+    this.registries = new ModuleRegistries(module =>
+      this.esmLoader.requireResultFromModule(module),
+    );
     invariant(
       this._environment.moduleMocker,
       '`moduleMocker` must be set on an environment when created',
@@ -221,6 +223,7 @@ export default class Runtime {
         this.requireModuleOrMock(from, moduleName),
       resolution: this._resolution,
       shouldLoadAsEsm: modulePath => this.unstable_shouldLoadAsEsm(modulePath),
+      testPath: this._testPath,
       testState: this.testState,
       transformCache: this.transformCache,
     });